// src/models/Wall.ts
export class Wall {
  number: number = 0;
  leftBottomNode: number = 0;
  rightBottomNode: number = 0;
  leftTopNode: number = 0;
  rightTopNode: number = 0;
  materialNumber: number = 0;
  method: number = 0;
  p1: number = 0;
  p2: number = 0;
  p3: number = 0;
  p4: number = 0;
  isShown: boolean = true;
}